requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
# streamlit==1.28.1
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
chromadb==0.4.22
pypdf==3.17.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-docx==1.1.0
python-multipart==0.0.6
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.36.0
torch==2.2.0
google-generativeai==0.3.2